# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Quiz - Teste Quiz EBC",
        "description": "API para gerenciamento de quizzes, participação de usuários e ranking.",
        "version": "0.3.0"
    },
    "paths": {
...
        "/api/quizzes/{quizId}/questions/{id}": {
            "get": {
                "tags": [
                    "Perguntas"
                ],
                "summary": "Exibe uma pergunta específica de um quiz",
                "description": "Retorna os detalhes de uma pergunta específica associada a um quiz.",
                "operationId": "showQuestion",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID do quiz ao qual a pergunta pertence",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pergunta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da pergunta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID da pergunta",
                                            "type": "integer"
                                        },
                                        "quiz_id": {
                                            "description": "ID do quiz",
                                            "type": "integer"
                                        },
                                        "text": {
                                            "description": "Texto da pergunta",
                                            "type": "string"
                                        },
                                        "correct_answer": {
                                            "description": "Resposta correta",
                                            "type": "boolean"
                                        },
                                        "image": {
                                            "description": "Caminho da imagem da pergunta",
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pergunta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Pergunta não encontrada. Verifique o ID e tente novamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao exibir a pergunta."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}